Extract failure descriptions into separate files
